Refactor StabilityAI and OpenAI code to use same ImagePrompt
* Simplify StabilityAI and OpenAI implementations to use a common shared ImagePrompt based on a similar 'Message' data type style used in ChatClient implementations
* Change StabilityAiImageOptions `seed` property type from `Integer` to `Long`
and update seed type to Long

TODOs
=====
Image.java should have
* getBytes()
* An enum that indicates the type of payload received, a url or a base64 or byte array.  Convert from url/base64 to byte array.
** No support yet in stability AI for returing directly a png.  AcceptHeader media/png
** need also an enum to indiate if the byte array or base64 is png,gif,jpg, etc.

+++ models/spring-ai-stabilityai/src/main/java/org/springframework/ai/stabilityai/StabilityAiImageGenerationMetadata.java
@@ -8,9 +8,9 @@ public class StabilityAiImageGenerationMetadata implements ImageGenerationMetada
 
 	private String finishReason;
 
-	private Integer seed;
+	private Long seed;
 
-	public StabilityAiImageGenerationMetadata(String finishReason, Integer seed) {
+	public StabilityAiImageGenerationMetadata(String finishReason, Long seed) {
 		this.finishReason = finishReason;
 		this.seed = seed;
 	}
@@ -19,7 +19,7 @@ public class StabilityAiImageGenerationMetadata implements ImageGenerationMetada
 		return finishReason;
 	}
 
-	public Integer getSeed() {
+	public Long getSeed() {
 		return seed;
 	}
