+++ models/spring-ai-stabilityai/src/main/java/org/springframework/ai/stabilityai/api/StabilityAiImageOptions.java
@@ -12,7 +12,7 @@ public interface StabilityAiImageOptions extends ImageOptions {
 
 	Integer getSamples();
 
-	Integer getSeed();
+	Long getSeed();
 
 	Integer getSteps();
