+++ models/spring-ai-stabilityai/src/main/java/org/springframework/ai/stabilityai/api/StabilityAiImageOptionsBuilder.java
@@ -52,7 +52,7 @@ public class StabilityAiImageOptionsBuilder {
 		return this;
 	}
 
-	public StabilityAiImageOptionsBuilder withSeed(Integer seed) {
+	public StabilityAiImageOptionsBuilder withSeed(Long seed) {
 		options.setSeed(seed);
 		return this;
 	}
